Various concurrency-related bug fixes:
* there has been a conflicting access to the owner property in class DOMFile: in some cases, a
second thread set the property to itself while another thread has been in the process of writing data.
As the owner object is used to determine the current page in the document, the writing thread
used a wrong data page (for a very short period). Thus, one or two document nodes got lost.

* a number of small caching problems in dom.dbx led to inconsistencies in the db. Also, queries
using string-equality comparisons did not use the cache, so increasing the cache size had no
positive effect on query speed.

svn path=/trunk/eXist-1.0/; revision=436


Former-commit-id: ee368109786d93e0255b0c3298fb3bbb22ab63d7