https://sourceforge.net/tracker/index.php?func=detail&aid=1019695&group_id=17691&atid=117691
org.exist.source.CocoonSource.getContent() throws NPE if the underlying InputSource cannot report its getContentLength(). For example, when initialized with an org.apache.excalibur.source.Source, which can legitimately report -1 for getContentLength() if the length is unknown a priori, then the CocoonSource.getContent() fails when trying to construct the outgoing ByteArrayOutputStream(source.getContentLength()).

This commit fixes the problem.

svn path=/trunk/eXist-1.0/; revision=589


Former-commit-id: 5dbbf40079161b6f689425674e5a7064f95de788