Applied patches by Gary Larsen:
-[1628975] Add additional info for StatusGenerator
Add status:cache-size and status:page-size as children to status:database-instance. This allows calculating the percentage of cache used which should be helpful in config tuning.

-[1628981] Cache usage info for StatusGenerator
Displays stats and graph of cache size usage. Also modified some graph colors so that buffer hits are green (green is good) and a glance can spot problems.

svn path=/trunk/eXist/; revision=5150


Former-commit-id: b470b50d0776ab372a133b0c1b03eade5c0755bf