New combined range index on named element or attribute. This mirrors recent changes on the fulltext index. In addition to the general range index, e.g. configured with
<create path="//author" type="xs:string"/>

one can now create an index on a given, fixed element or attribute qname:

<create qname="author" type="xs:string"/>

The existance of this index allows eXist to apply better optimization techniques, reducing the cost for a query like //book[author = 'John Doe"] to a single index lookup. However, eXist will currently not apply the new optimization methods automatically, so - for the time being - you have to give it a hint:

//((#exist:optimize#) { book[author = "John Doe"] })

For more examples see test/src/org/exist/performance/test.xml. The plan is to have the XQuery compiler insert those optimization hints automatically when analyzing the query. Right now this has to be done manually though.

svn path=/trunk/eXist/; revision=5288


Former-commit-id: 7545673b0ce7b13d693ee2997e17326557d49d41