Final piece of custom protocol handler.
Now it is possible to store grammar files and catalog files inside the database. For this, change conf.xml like:

        <entity-resolver>
            <catalog uri="xmldb:exist:///db/mygrammar/catalog.xml" />
        </entity-resolver>

please note that when a file is being stored, this should not be done into the collection in which a catalog files are stored and where the grammar files are stored. We need to fix this  here, but also for the webDAV code. Same potential bugs.

for syntax of catalog 1.1 files:
http://www.oasis-open.org/committees/download.php/14809/xml-catalogs.html

svn path=/trunk/eXist/; revision=5750


Former-commit-id: be99008a2631fab3de95e9da59137cc6262521ae