Changed FileRead function to use the custom UnicodeReader instead of InputStreamReader, to work around known bug in InputStreamReader not handling the Unicode BOM (Byte Order Mark) for UTF-8 encoded files.
This bug would manifest if you tried to do something like util:parse( util:file-read( "myfile.xml" ) ) in an XQuery.  If the file included the valid UTF-8 BOM characters ( xEFBBBF ) at the start, then they would be included in the string returned from the file-read and thus the parse would fail with a "characters in prolog" xml parsing error, which is not correct behaviour for a standard UTF-8 file.  See comments in UnicodeReader source for more information.

svn path=/trunk/eXist/; revision=7488


Former-commit-id: f9cd85433665a8e96fa4ea329a388180d5f04749