Performance:  External modules (i.e., ones defined in XQuery) are now cached in the XQueryPool, and can be shared across cached queries.  This greatly improves performance if many queries are importing the same module since the module only has to be compiled once.  The code is a little hairy, but the unit tests and XQTS are no worse off than before (and the module tests in XQTS seem to be running a lot faster).
Performance:  Took out eager initialization of the calendar and time zone in the query context, since getting the calendar is surprisingly slow.  They'll be initialized lazily on demand.  This could make a slight difference in long-running queries that don't try to get the current time until later, but is spec-compliant.

svn path=/trunk/eXist/; revision=8417


Former-commit-id: 01429fd3ac2dce59c011de74fb7a11e0896b0673