[feature] Ant eXist 'store' task has some limitations:
1.- It is not possible to tell the mime-types.xml file to use when MIME
types and xml/binary resource type are guessed.

2.- When 'type' attribute is set to "xml" or "binary", MIME type
associated to the file to be stored is lost, acquiring application/xml when
it is an xml resource, and application/octet-stream when it is a binary
resource. Something desirable would be maintaining the found MIME type, but
marking the resource according to 'type' attribute.

3.- It is not possible to set a default custom MIME type to the
resources to store when they are unknown, or to force the MIME type.

So these changes overcome those limitations, adding some attributes to this
Ant eXist task.


svn path=/trunk/eXist/; revision=9149


Former-commit-id: 44bd12a980026ce0bcf51feab3a95ed535f9c46d