Changed range index lookups so that they will use collations for starts-with(), ends-with() and contains() and the newly added equals() functions in predicates.
Where this is useful, is that you can specify something like:

 	$element[ starts-with( @attribute, "abc", "?strength='secondary' ) ]

and the starts-with will match regardless of the case of @attribute.  Eg. it will match "abcdef" or "ABCXYZ" in the @attribute value, using the range index and not single stepping the predicate comparisons.

As part of this new functionality, a serious bug found in all three of the above functions was rectified. Prior to this, the 2nd argument string was interpreted as a regex match string!  So if you specified (or have a node with a value) of something like ""^*", for the 2nd function parameter,  then the functions would always return true. Ouch!  This is no longer the case, as the 2nd string parameter is now used as a literal string value and not a regex match.

For example, prior to this fix,

	starts-with( "abc", "^*" ) = true, which is not correct according to the specs AFAIK.
	starts-with( "abc", "*" )  threw an exception with an invalid RegEx pattern message.

Now, with the fix:

	starts-with( "abc", "^*" ) = false 
	starts-with( "abc", "*" )  = false

which is the correct behaviour.

Also added a new fn:equals() function, which basically is the same as the "=" operation, except that it uses collations for range indexes, and like the other three functions, it can take a custom collation as an optional third parameter.


	


svn path=/trunk/eXist/; revision=10673


Former-commit-id: bc312fcb045d854e1595aecb3aadddb7bf6969e8