[feature] Adding stubs for new Subversion functions.
svn:clean-up($repository-uri as xs:anyURI, $username as xs:string, $password as xs:string, $resource as xs:anyURI, $delete as xs:boolean) empty()

? a resource to a subversion repository.

This is a stub and currently does nothing.


$repository-uri		The location in the subversion repository URI
$username		The subversion username
$password		The subversion password
$resource		The path to the resource.

 ========================

svn:commit($repository-uri as xs:anyURI, $username as xs:string, $password as xs:string, $resource as xs:anyURI, $message as xs:string?) node()

Commits a resource to a subversion repository.

This is a stub and currently does nothing.


$repository-uri		The location in the subversion repository URI from which the logs should be retrieved
$username		The subversion username
$password		The subversion password
$resource		The path to the resource to be stored.
$message		The SVN commit message.

Returns The commit information.

 ========================

svn:lock($repository-uri as xs:anyURI, $username as xs:string, $password as xs:string, $resource as xs:anyURI, $message as xs:string) empty()

Locks a resource to a subversion repository.

This is a stub and currently does nothing.


$repository-uri		The location in the subversion repository URI
$username		The subversion username
$password		The subversion password
$resource		The path to the resource.
$message		The lock message

 ========================

svn:unlock($repository-uri as xs:anyURI, $username as xs:string, $password as xs:string, $resource as xs:anyURI) empty()

Unlocks a resource in a subversion repository.

This is a stub and currently does nothing.


$repository-uri		The location in the subversion repository URI
$username		The subversion username
$password		The subversion password
$resource		The path to the resource.

 ========================

svn:update($repository-uri as xs:anyURI, $username as xs:string, $password as xs:string, $resource as xs:anyURI, $revision as xs:integer?) xs:long

Updates a resource from a subversion repository.

This is a stub and currently does nothing.


$repository-uri		The location in the subversion repository URI from which the logs should be retrieved
$username		The subversion username
$password		The subversion password
$resource		The path to the resource to be stored.
$revision		The revision number to update to. An empty value updates to the latest revision.

Returns The commit information.



svn path=/trunk/eXist/; revision=11725


Former-commit-id: 531fc740732671a590d3752b43f5e4a20dc3e5f3