Added new attribute to db-connection in conf.xml.tmpl:  checkMaxCacheSize=true.
This value can be used to disable checking of the max cache size for users with tons of memory allocated to their JVM, typically 64-bit JVM users on machines with gobs of memory.

Suitable caveats are included in the comments, and a warning is written to exist.log if max cache size checking is disabled.

Default value is true.

svn path=/trunk/eXist/; revision=11990


Former-commit-id: a5390d6ed1417ff82669c7612a3544bdfa59f384