Added the ability to override configuration attribute values specified in conf.xml using system properties.
Any unique attributes specified in conf.xml can be overriden using a Java system property, typically specified on the command line, of the form:
    
    org.element.element....attribute
    	
where the nesting of the element names follows the structure of the XML element heirarchy found in conf.xml.
    	
    For example, to override the value of the cache size to be 128MB you would specify:
    
    	-Dorg.exist.db-connection.cacheSize="128M"
    	
on your JVM startup command line or options.  Note that this only works for unique, non-repeating elements, so you can't override things like the transformer attribute element values or the XQuery module builtin definitions, since they are not unique.

This is a very handy feature if you have values you want to test, but don't want to have to reconfigure eXist for some transient testing.  It is also useful if you deploy eXist using a war file, but might have multiple different databases in different locations.  You can have a single war file with a common conf.xml, but change which database location you want to use, at startup, by specifying a JVM property such as:

    -Dorg.exist.db-connection.database=/home/me/my/custom/database/dir

If a configuration attribute value is overriden using this mechanism, a warning is written to the exist.log file highlighting which value was overriden, so that it's clear where the configuration parameter came from.

svn path=/trunk/eXist/; revision=12055


Former-commit-id: bf8dcf69788fa1c23d48d7c63eda1937ba941bab